Implemented getRelease function in the titles controller and change releases route prefix.
