+++ server/src/routers/titles.js
@@ -1,10 +1,10 @@
 import { Router } from "express";
-import { getReleases } from "../controllers/movie.js";
+import { getReleases } from "../controllers/titles.js";
 
 
 const router = Router();
 
 router.get("/releases", getReleases);
 
-export { router as movieRouter };
+export { router as titleRouter };
